refactor: altera logica para receber expiresAt como milesegundos
