+++ frontend/src/services/login-service.ts
@@ -4,7 +4,7 @@ import { apiFetch } from '@/lib/api-client';
 import { getApiUrl } from '@/lib/api-url';
 import { LoginForm } from '@/types/login-schema';
 import { ServerDTO } from '@/types/server-dto';
-import { AuthResponse } from '@/types/user-schema';
+import { AuthResponse } from '@/types/auth-response';
 
 const localhost = getApiUrl();
