+++ frontend/src/types/movements-report-schema.ts
@@ -1,6 +1,6 @@
 import { Product } from './product-schema';
 import { StockLocationsData } from './stock-location-schema';
-import { User } from './user-schema';
+import { User } from './auth-response';
 
 export type MovementsReportParams = {
   search: string;
